Added a new page that displays shelf's details
Now there's a new shelf_details.html that defines how the page that contains shelf's information will be displayed.
Added a new route in the backend that manages info extraction in order to give them to the shelf details page.
Modified photoTransfer code in the camera. Now for the webserver IP uses the same specified in the config struct configured by the user