Implemented the new "Add Shelf" feature
Now it's possible to add a new shelf. For each new shelf we have to set a number (could contains letters as well) a name.
We can link cameras registered by the backend that are not associated to a shelf yet.
For doing so, new HTML and JS files are added and two new routes in the Flask webapp are inserted (one for access to the "Add Shelf" page and one for doing the operation)